third_party/blink/web_tests/platform/linux/external/wpt/pointerevents/pointerevent_pointercapture_in_frame_pen-expected.txt

This is a testharness.js-based test.
PASS Test penpointer capture in same-origin frame: Pointer down at inner frame and set pointer capture.
FAIL Test penpointer capture in same-origin frame: Pointer down at outer frame body and set pointer capture. assert_equals: Expected pointerup event, but got lostpointercapture event instead expected "pointerup" but got "lostpointercapture"
PASS Test penpointer capture in same-origin frame: Pointerdown with set capture at inner frame, then release on next pointermove.
PASS Test penpointer capture in same-origin frame: Pointerdown with set capture at outer frame, then release on next pointermove.
Harness: the test ran to completion.
